multi command comment component out of Explore
